create base class
