+++ csrc/scheduler/ampere_multi_matmul.h
@@ -9,6 +9,7 @@
 
 #include <ATen/cuda/CUDAContext.h>
 #include <scheduler/mma_utils.h>
+#include <scheduler/multi_matmul.h>
 #include <val_graph.h>
 #include <val_graph_visitor.h>
 
@@ -66,13 +67,11 @@ namespace nvfuser {
 // Each of the named tensors above is scheduled differently. We schedule them
 // by building AbstractTensors for each tensor category; these are held in
 // AmpereMultipleMatmulScheduler::schedules_.
-// TODO: Inheret from SchedulerEntry
-class AmpereMultipleMatmulScheduler {
+// TODO: Inherit from SchedulerEntry
+class AmpereMultipleMatmulScheduler : public MultipleMatmulScheduler {
  public:
   AmpereMultipleMatmulScheduler(Fusion* fusion, const MatmulParams* params)
-      : fusion_(fusion),
-        params_(params),
-        id_model_(fusion, /*build_graphs=*/false) {
+      : MultipleMatmulScheduler(fusion, params) {
     const auto device_prop = at::cuda::getCurrentDeviceProperties();
     const int cc = device_prop->major * 10 + device_prop->minor;
     NVF_ERROR(
@@ -80,22 +79,10 @@ class AmpereMultipleMatmulScheduler {
         "This matmul scheduler is restricted to Ampere and Turing.");
   }
 
-  void run();
+  void run() final;
 
  private:
   void cacheInputsAndOutputs();
-
-  void findPatterns();
-
-  void countDims();
-
-  void translatePatterns();
-
-  // Get tensor roles and id roles
-  // When there are multiple matmul patterns, we can have conflicting roles.
-  // For now we throw an error if this is the case.
-  // TODO: This should be checked in canScheduleCompileTime
-  void findRoles();
 
   // Including current tensor naming convention for reference,
   //  this is very temporary and will change over time and
@@ -147,11 +134,6 @@ class AmpereMultipleMatmulScheduler {
   void addSetsForCacheReads(
       const std::vector<TensorView*>& tv_smems,
       std::vector<TensorView*>& tv_rs);
-
-  //! Rebuilds IdModel, then updates all ValGroups in abstract tensors to refer
-  //! to the new IdModel. This is necessary whenever we perform an operation
-  //! that creates a new TensorView, such as caching or rFactor
-  void updateIdModel();
 
   //! Swizzle the M and N outer dimensions after makeTile has been called.
   //! This updates outer_dim_roles if we introduce a new dimension, which can
@@ -216,26 +198,12 @@ class AmpereMultipleMatmulScheduler {
   void setUpCircularBuffering();
 
  private:
-  Fusion* fusion_;
-  const MatmulParams* params_;
-  IdModel id_model_;
-  // Permissive graph of id_model_, which we modify at times using e.g.
-  // AbstractTensor.split or by mapping vals in cacheAfter and rFactor
-  ValGraph* graph_ = nullptr;
-  std::vector<mma_utils::MatmulPattern> patterns_;
-  mma_utils::DimRolesMap id_roles_;
-  mma_utils::TensorRolesMap tensor_roles_;
-  mma_utils::MatmulOperandInnerDims inner_dims_;
-
-  int64_t num_splitk_dims_ = 0, num_device_dims_ = 0, num_local_batch_dims_ = 0,
-          num_device_and_batch_dims_ = 0;
-
   std::vector<std::pair<TensorView*, TensorView*>> cached_outputs_;
 
   std::vector<ValGroup> canonical_dim_ordering_;
 
-  std::vector<TensorView*> as_, bs_, acw_smems_, bcw_smems_, acrs_, bcrs_, abs_,
-      bbs_, mma_results_, splitk_sums_, smem_epilogues_;
+  std::vector<TensorView*> acw_smems_, bcw_smems_, acrs_, bcrs_, abs_, bbs_,
+      splitk_sums_, smem_epilogues_;
 };
 
 } // namespace nvfuser
