+++ csrc/scheduler/mma_utils.h
@@ -463,7 +463,17 @@ std::optional<std::pair<DimRolesMap, TensorRolesMap>> allPatternRoles(
     const std::vector<MatmulPattern>& patterns);
 
 // Utility to check concrete static size
-inline void checkConcreteStaticDim(const AbstractId& abs_id);
+inline void checkConcreteStaticDim(const AbstractId& abs_id) {
+  IterDomain* id = representativeId(abs_id);
+  NVF_ERROR(
+      !id->isBroadcast() && !id->isReduction(),
+      "no support for reduction or broadcast domains, but got ",
+      id->toString());
+  NVF_ERROR(
+      id->extent()->isConstInt(),
+      "swizzled dimension's extend must be known during scheduling, got ",
+      id->toString());
+}
 
 } // namespace mma_utils
